Migrate Mutation.updateOrganizationSlug
