+++ packages/services/api/src/modules/organization/resolvers.ts
@@ -14,14 +14,6 @@ import { Logger } from '../shared/providers/logger';
 import type { OrganizationModule } from './__generated__/types';
 import { OrganizationManager } from './providers/organization-manager';
 
-const OrganizationSlugModel = z
-  .string({
-    required_error: 'Organization slug is required',
-  })
-  .min(1, 'Organization slug is required')
-  .max(50, 'Slug must be less than 50 characters')
-  .regex(/^[a-z0-9-]+$/, 'Slug can only contain lowercase letters, numbers and dashes');
-
 const createOrUpdateMemberRoleInputSchema = z.object({
   name: z
     .string({
@@ -46,45 +38,6 @@ const createOrUpdateMemberRoleInputSchema = z.object({
 
 export const resolvers: OrganizationModule.Resolvers = {
   Mutation: {
-    async updateOrganizationSlug(_, { input }, { injector }) {
-      const parsedInput = OrganizationSlugModel.safeParse(input.slug.trim());
-
-      if (!parsedInput.success) {
-        return {
-          error: {
-            message:
-              parsedInput.error.formErrors.fieldErrors?.[0]?.[0] ??
-              'Changing the organization slug failed.',
-          },
-        };
-      }
-
-      const organizationId = await injector.get(IdTranslator).translateOrganizationId(input);
-      const result = await injector.get(OrganizationManager).updateSlug({
-        slug: input.slug,
-        organization: organizationId,
-      });
-
-      if (result.ok) {
-        return {
-          ok: {
-            updatedOrganizationPayload: {
-              selector: {
-                organization: result.organization.cleanId,
-              },
-              organization: result.organization,
-            },
-          },
-        };
-      }
-
-      return {
-        ok: null,
-        error: {
-          message: result.message,
-        },
-      };
-    },
     async joinOrganization(_, { code }, { injector }) {
       const organization = await injector.get(OrganizationManager).joinOrganization({ code });
